Enhance Makefile and update dependencies for cdot65.scm Ansible Collection
This commit modifies the Makefile to include additional Tox environments for linting and formatting, improving the development workflow. It also updates the dependencies in `pyproject.toml` and `poetry.lock`, upgrading `ansible-core` to version 2.18.0 and `pan-scm-sdk` to version 0.3.44. Furthermore, the README.md is updated to reflect the new Python and Ansible version requirements, ensuring clarity for users. These changes enhance the project's maintainability and align it with the latest standardpdate